+++ src/BadaBing.vala
@@ -26,6 +26,7 @@ namespace BadaBing {
     public enum Desktop {
         Gnome,
         Ubuntu,
+        Pantheon,
         Pop,
         Mate,
         PCManFM,
@@ -92,7 +93,7 @@ Keywords=Wallpaper;
 
     public const string CRONJOB_DIR     = ".config/badabing";
     public const string CRONJOB_PATH    = ".config/badabing/cronjob.sh";
-    public const string CATLOCK_PATH    = ".local/share/catlock/themes/badabing/copy.sh";
+    public const string CATLOCK_PATH    = ".local/share/catlock/themes/badabing.copy.sh";
 
     public const int BUFLEN = 256;